add support for PDFlib 7/8/9
git-svn-id: http://svn.php.net/repository/pecl/pdf/trunk@330855 c90b9560-bf6c-de11-be94-00142212c4b1

+++ config.m4
@@ -6,12 +6,13 @@ PHP_ARG_WITH(pdflib,for PDFlib support,
 [  --with-pdflib[=DIR]     Include PDFlib support.])
 
 if test "$PHP_PDFLIB" != "no"; then
-
   dnl #
   dnl # The main PDFlib configure
   dnl #
 
-  PHP_NEW_EXTENSION(pdf, pdf.c, $ext_shared)
+  PHP_REQUIRE_CXX()
+  dnl # $6 set to "1" => use $(CXX) for linking
+  PHP_NEW_EXTENSION(pdf, pdf.c, $ext_shared, "", "", 1)
   dnl # MacOSX requires this
   case `(uname -s) 2>/dev/null || echo unknown` in
     *arwin*)
@@ -23,36 +24,22 @@ if test "$PHP_PDFLIB" != "no"; then
 
   case $PHP_PDFLIB in
     yes)
-      AC_CHECK_LIB(pdf, PDF_open_pdi, [
         AC_DEFINE(HAVE_PDFLIB,1,[ ])
         PHP_ADD_LIBRARY(pdf,, PDF_SHARED_LIBADD)
-      ],[
-        AC_MSG_ERROR([
-PDFlib extension requires at least pdflib 4.0.x.
-See config.log for more information.
-])
-      ],[
-          -lm $PHP_FRAMEWORKS
-      ])
     ;;
     *)
       if test -f "$PHP_PDFLIB/include/pdflib.h" ; then
-
-        PHP_CHECK_LIBRARY(pdf, PDF_open_pdi, 
-        [
           AC_DEFINE(HAVE_PDFLIB,1,[ ]) 
           PHP_ADD_LIBRARY_WITH_PATH(pdf, $PHP_PDFLIB/lib, PDF_SHARED_LIBADD)
           PHP_ADD_INCLUDE($PHP_PDFLIB/include)
-        ],[
-          AC_MSG_ERROR([
-PDFlib extension requires at least pdflib 4.0.x.
-See config.log for more information.
-])
-        ],[
-          -L$PHP_PDFLIB/lib $PDF_SHARED_LIBADD -lm
-        ])
       else
-        AC_MSG_ERROR([pdflib.h not found! Check the path passed to --with-pdflib=<PATH>. PATH should be the install prefix directory.])
+	  if test -f "$PHP_PDFLIB/pdflib.h" ; then
+	      AC_DEFINE(HAVE_PDFLIB,1,[ ]) 
+	      PHP_ADD_LIBRARY_WITH_PATH(s_libpdf, $PHP_PDFLIB, PDF_SHARED_LIBADD)
+	      PHP_ADD_INCLUDE($PHP_PDFLIB)
+	  else
+	      AC_MSG_ERROR([pdflib.h not found! Check the path passed to --with-pdflib=<PATH>. PATH should be the install prefix directory.])
+	  fi
       fi
     ;;
   esac
